pyproject.toml
[project]
name = "daily-trading-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.2.0",
    "joblib>=1.5.3",
    "lightgbm>=4.6.0",
    "matplotlib>=3.10.8",
    "nbformat>=5.10.4",
    "optuna>=4.7.0",
    "pandas<3.0.0",
    "plotly>=6.6.0",
    "polars>=1.38.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "simfin>=1.0.1",
    "streamlit>=1.40.0",
    "xgboost>=3.2.0",
]
